Phaser debug mode and test sprite
I added in the code to turn phaser's debug mode on and then turned it on. It outlines all sprites in a red box so you can see where they are on the screen even if they are not being displayed properly and also provides a blue line pointing in the direction you are supposed to be moving if phaser is reading that you should be moving.

I also replaced the little red dot with an actual test sprite!